Melhorando verificação de entrada de dados e consistencia para <No Content>

+++ app/controller/ControladorApp.class.php
@@ -34,12 +34,25 @@ class ControladorApp
         $status = 200;
 
         try {
-            $dao = new BoloDAO();
-            $bolosArray = $dao->getBoloById( $args["idBolo"] );
 
-            $corpoResp =  json_encode( $bolosArray );
+            if ( isset( $args["idBolo"]) ) {
+                if ( !is_numeric( $args["idBolo"] ) )
+                    throw new BadHttpRequest();
+            } else { 
+                throw new BadHttpRequest();
+            }
+
+            $dao = new BoloDAO();
+            $bolo = $dao->getBoloById( $args["idBolo"] );
+
+            $status = ( !is_null( $bolo ) ) ? 200 : 204; 
+
+            $corpoResp =  json_encode( $bolo );
             $response = $response->withHeader('Content-type', 'application/json')
                                  ->write( $corpoResp );
+        } catch (BadHttpRequest $e) {
+            $status = 400;
+            $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
         } catch ( \PDOException $e ) {
             $status = 500;
             $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
@@ -52,12 +65,17 @@ class ControladorApp
         $status = 200;
 
         try {
+            if ( !isset( $args["idBolo"] ) )
+               throw new BadHttpRequest();
+
             $dao = new BoloDAO();
+            # TRUE: Deletou a tupla com o id requisitado => 200
+            # FALSE: Não existia nenhuma tupla com o id requisitado => 204
             $sucesso = $dao->deleteBoloById( $args["idBolo"] );
-
-            var_dump($sucesso);
-
-            $status = ( $sucesso ) ? $status = 200 : $status = 204;
+            $status = ( $sucesso ) ? 200 : 204;
+        } catch (BadHttpRequest $e) {
+            $status = 400;
+            $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
         } catch ( \PDOException $e ) {
             $status = 500;
             $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
@@ -106,4 +124,48 @@ class ControladorApp
 
         return $response->withStatus($status);
     }
+
+    public function updateBolo( Request $request, Response $response, array $args )
+    {
+        # Verificar erro na entrada
+        # Esperamos uma entrada Json da seguinte forma:
+        # {"nome":"valor","sabor":"valor","cobertura":"valor","descricao":"valor"}
+        $objEntrada = $request->getParsedBody();
+        $status = 201;
+
+        try {
+            # getParsedBody, irá parsear o Json de entrada, caso ele falhe
+            # pela entrada não ser um Json válido, esse if irá capturar
+            if ( is_null($objEntrada) )
+                throw new BadHttpRequest();
+
+            # Verifica se os campos obrigatórios estão preenchidos
+            if (!( isset( $objEntrada["nome"] ) &&
+            isset( $objEntrada["sabor"] ) &&
+            isset( $objEntrada["cobertura"])))
+                throw new BadHttpRequest();
+
+            # Descrição é opcional, então caso não tenha sido passada será substituida por ""
+            $desc = (!\is_null($objEntrada["descricao"])) ? $objEntrada["descricao"] : "";
+
+            $arrayBolo = array( "nome"=>$objEntrada["nome"],
+                                "sabor"=>$objEntrada["sabor"],
+                                "cobertura"=>$objEntrada["cobertura"],
+                                "descricao"=> $desc);
+
+            $boloInst = new Bolo($arrayBolo);
+            $dao = new BoloDAO();
+            $dao->updateBolo( 1 , $boloInst );
+        } catch (BadHttpRequest $e) {
+            $status = 400;
+            $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
+        } catch (\PDOException $e) {
+            $status = 500;
+            $response->write('Exceção capturada: '.  $e->getMessage(). '\n');
+        }
+
+        return $response->withStatus($status);
+    }
+
+
 }
